requirements.txt
Flask==3.1.2
spotipy==2.25.1
python-dotenv==1.1.1